# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  apperrors.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Cart:
    properties:
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.CartItem:
    properties:
      cart_id:
        type: integer
      id:
        type: integer
      quantity:
        type: integer
      smartphone_id:
        type: integer
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.SetLang:
    properties:
      lang:
        example: ru
        type: string
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
  models.UpdateRequest:
    properties:
      avatar:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.User:
    properties:
      avatar:
        type: string
      cart:
        $ref: '#/definitions/models.Cart'
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API Server for Smartbuy application
  termsOfService: http://swagger.io/terms/
  title: Smartbuy API
  version: "1.0"
paths:
...
  /language:
    post:
      consumes:
      - application/json
      description: Saves the user's preferred language (key="lang") in a cookie for
        30 days.
      parameters:
      - description: Language (e.g. ru, en)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SetLang'
      produces:
      - application/json
      responses:
        "201":
          description: New Language
          schema:
            $ref: '#/definitions/models.SetLang'
        "400":
          description: Invalid language code
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Set Interface Language
      tags:
      - settings
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and signs a JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: email&password
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: User Login
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Creates a new user account and sends a tmp password (single use)
        to the user's email
      parameters:
      - description: Name and Email of the user
        in: body
        name: name&email
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      summary: Register User
      tags:
      - auth
...
  /users:
    get:
      description: Admin only. Lists all registered users.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - users
  /users/{user_id}:
    delete:
      description: Permanently remove a user account. Users can delete themselves;
        Admins can delete anyone.
      parameters:
      - description: User ID to delete
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns the deleted user data
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "403":
          description: Forbidden - Not your account
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete User Account
      tags:
      - users
    get:
      description: Get details of a specific user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update name, avatar, or password.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Update Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"